Fix identifiers collision in subreport list items
Mixing SubReporList classes of distinct reports (companies, domains,
repositories, etc) in a sigle page produced unexpected errors when
two or more items had the same name (i.e, the company "Bitergia",
and the domain "bitergia"). This commit adds a prefix to the id
of the items, based on the type of report, that prevents from name
collision.